+++ app/src/main/java/kr/sparta/tripmate/ui/budget/budgetList/BudgetFragment.kt
@@ -1,4 +1,4 @@
-package kr.sparta.tripmate.ui.budget
+package kr.sparta.tripmate.ui.budget.budgetList
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -11,8 +11,8 @@ import kr.sparta.tripmate.data.model.budget.Budget
 import kr.sparta.tripmate.databinding.FragmentBudgetBinding
 import kr.sparta.tripmate.ui.budget.budgetcontent.BudgetContentActivity
 import kr.sparta.tripmate.ui.budget.budgetdetail.main.BudgetDetailActivity
-import kr.sparta.tripmate.ui.viewmodel.budget.BudgetViewModel
-import kr.sparta.tripmate.ui.viewmodel.budget.BudgetFactory
+import kr.sparta.tripmate.ui.viewmodel.budget.budgetlist.BudgetViewModel
+import kr.sparta.tripmate.ui.viewmodel.budget.budgetlist.BudgetFactory
 
 class BudgetFragment : Fragment() {
     companion object {
