+++ app/src/main/java/kr/sparta/tripmate/ui/main/ViewPager2Adapter.kt
@@ -4,7 +4,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import kr.sparta.tripmate.R
-import kr.sparta.tripmate.ui.budget.BudgetFragment
+import kr.sparta.tripmate.ui.budget.budgetList.BudgetFragment
 import kr.sparta.tripmate.ui.community.main.CommunityFragment
 import kr.sparta.tripmate.ui.home.HomeFragment
 import kr.sparta.tripmate.ui.mypage.home.MyPageFragment
