+++ app/src/main/res/layout/fragment_budget.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.budget.BudgetFragment">
+    tools:context=".ui.budget.budgetList.BudgetFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/budget_list_toolbar"
